Issue with parallel downloading
I found that this code was failing for a wide variety of inputs. Here's why (for me):

pool_search.join() was being annihilated for experiments with many files, and hitting IOError: [Errno 24] Too many open files. 

E.g., there are something like that 1000 individual siconc files for EC-Earth3. I didn't see the rationale in having pool_search around, as this was much less time-intensive than pool_download. So I removed it as get_files_to_download will append files_to_download on its own, and this operation is quite simple. 

Also, pool_download was being assigned the maximum of # processors and # files, not the minimum. It would request several hundred workers and crash. 

As edited this seems to work for me on my local machine (with edits to outpath, etc).